requirements.txt
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.13
openai>=0.28.1
python-dotenv>=1.0.0
customtkinter>=5.2.0
Pillow>=10.0.0
